requirements.txt
fake_useragent==2.0.3
fastapi==0.115.8
imagebind==0.1.0
numpy==1.25.2
openai==1.61.0
pandas==2.0.0
pgvector==0.3.6
pillow==10.4.0
psycopg2_binary==2.9.10
pydantic==2.10.6
python-dotenv==1.0.1
Requests==2.32.3
rich==13.7.1
SQLAlchemy==2.0.36
starlette==0.45.3
torch==2.3.0
tqdm==4.67.1
transformers==4.44.2
flash-attn==2.5.8
faiss-gpu==1.7.2
beir==2.0.0
mteb==1.29.0